singlepointmeasurement comparator

+++ src/xerus/fullTensor.cpp
@@ -406,7 +406,7 @@ namespace xerus {
 		data.reset(newData, &internal::array_deleter_vt);
 		
 		// Adjust dimensions
-		dimensions.erase(dimensions.begin()+_dimension);
+		dimensions.erase(dimensions.begin()+(long)_dimension);
 		size = stepCount*blockSize;
 	}
